+++ yajl-1.0.11/yajl_buf.c
@@ -33,7 +33,6 @@
 #include "yajl_buf.h"
 
 #include <assert.h>
-#include <stdlib.h>
 #include <string.h>
 
 #define YAJL_BUF_INIT_SIZE 2048
@@ -46,7 +45,7 @@ struct yajl_buf_t {
 };
 
 static
-void yajl_buf_ensure_available(yajl_buf buf, unsigned int want)
+void yajl_buf_ensure_available(yajl_buf buf, size_t want)
 {
     unsigned int need;
     
@@ -84,7 +83,7 @@ void yajl_buf_free(yajl_buf buf)
     YA_FREE(buf->alloc, buf);
 }
 
-void yajl_buf_append(yajl_buf buf, const void * data, unsigned int len)
+void yajl_buf_append(yajl_buf buf, const void * data, size_t len)
 {
     yajl_buf_ensure_available(buf, len);
     if (len > 0) {
